Aariz/requirements.txt
# Core dependencies
torch>=1.12.0
torchvision>=0.13.0

# Data processing
numpy>=1.21.0,<2.0.0
Pillow>=9.0.0
pandas>=1.3.0

# Image processing
opencv-python>=4.5.0
scikit-image>=0.19.0
albumentations>=1.1.0

# Scientific computing
scipy>=1.7.0

# Utilities
tqdm>=4.62.0
tensorboard>=2.8.0
matplotlib>=3.5.0


